+++ sync-for-commerce/Codat/Sync/Commerce/Models/Components/Company.cs
@@ -112,10 +112,25 @@ namespace Codat.Sync.Commerce.Models.Components
         public string? CreatedByUserName { get; set; } = null;
 
         /// <summary>
+        /// An array of products that are currently enabled for the company.
+        /// </summary>
+        [JsonProperty("products")]
+        public List<string>? Products { get; set; }
+
+        /// <summary>
         /// A collection of user-defined key-value pairs that store custom metadata against the company.
         /// </summary>
         [JsonProperty("tags")]
-        public Tags? Tags { get; set; }
+        public Dictionary<string, string>? Tags { get; set; }
+
+        [JsonProperty("referenceParentCompany")]
+        public CompanyReference? ReferenceParentCompany { get; set; }
+
+        /// <summary>
+        /// A list of subsidiary companies owned or controlled by this entity. Empty if the company has no children.
+        /// </summary>
+        [JsonProperty("referenceSubsidiaryCompanies")]
+        public List<CompanyReference>? ReferenceSubsidiaryCompanies { get; set; }
 
         [JsonProperty("dataConnections")]
         public List<Connection>? DataConnections { get; set; }