+++ sync-for-commerce/RELEASES.md
@@ -113,3 +113,13 @@ Based on:
 - [csharp v4.0.0] sync-for-commerce
 ### Releases
 - [NuGet v4.0.0] https://www.nuget.org/packages/Codat.Sync.Commerce/4.0.0 - sync-for-commerce
+
+## 2024-11-26 16:28:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.447.0 (2.463.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v5.0.0] sync-for-commerce
+### Releases
+- [NuGet v5.0.0] https://www.nuget.org/packages/Codat.Sync.Commerce/5.0.0 - sync-for-commerce